+++ Source/GoBot/GoBotInternal.swift
@@ -206,9 +206,8 @@ class GoBotInternal {
     // MARK: planetary services
 
     // TODO: deprecated
-    private lazy var notifyNewBearerToken: CPlanetaryBearerTokenCallback = { cstr, expires in
-        return
-    }
+    private lazy var notifyNewBearerToken: CPlanetaryBearerTokenCallback = { _, _ in
+        }
 
     // MARK: connections
 
@@ -670,7 +669,7 @@ class GoBotInternal {
     
     // MARK: Aliases
     
-    func register(alias: String, in room: Room) -> Bool {
+    func register(alias: String, in room: Room) throws -> String {
         Log.debug("Registering room alias: \(alias) at \(room.address.string)")
         let result: ssbRoomsAliasRegisterReturn_t = room.address.string.withGoString { roomAddress in
             alias.withGoString { alias in
@@ -681,14 +680,18 @@ class GoBotInternal {
         if result.alias != nil {
             let aliasURLString = String(cString: result.alias)
             if aliasURLString.isEmpty == false {
-                // TODO: return alias or error
-                return true
+                return aliasURLString
             }
         }
-            return false
+        
+        switch result.err {
+        case 2:
+            throw RoomRegistrationError.aliasTaken
+        default:
+            throw RoomRegistrationError.unknownError
+        }
     }
     
     func revoke(alias: RoomAlias) async throws {
-        
     }
 }