+++ Source/UI/AvatarStackView.swift
@@ -59,6 +59,7 @@ class AvatarStackView: UIView {
         }
     }
 
+    @available(*, unavailable)
     required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }