+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.3.10] 2022-11-25
 
+- Add rooms and alias setup onboarding screen. #940
 - Disabled EBT when the user is restoring their feed to work around a bug in our pubs
 - Added new redesigned Profile screen. #681 #976 
 - Fixed a bug in macOS that prevented users from load the identity from the keychain. #978